requirements.txt
allure_python_commons==2.13.4
Faker==24.3.0
pytest==8.0.0
pytest==8.1.1
Requests==2.31.0
selenium==4.19.0
webdriver_manager==4.0.1
